+++ pages/update_data.php
@@ -56,6 +56,34 @@ if($_POST['value'] && $_GET['formula'] && $_POST['pk'] && !$_GET['settings']){
 	$name = mysqli_real_escape_string($conn, $_POST['name']);
 
 	mysqli_query($conn, "UPDATE ingSuppliers SET $name = '$value' WHERE id = '$sup_id'");
+	return;
+	
+}elseif($_GET['supp'] == 'add'){
+	$description = mysqli_real_escape_string($conn, $_GET['description']);
+	$name = mysqli_real_escape_string($conn, $_GET['name']);
+	
+	if(empty($name)){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error: </strong> Supplier name required</div>';
+		return;
+	}
+	if(mysqli_num_rows(mysqli_query($conn, "SELECT name FROM ingSuppliers WHERE name = '$name'"))){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>'.$name.'</strong> Supplier already exists!</div>';
+		return;
+	}
+
+	if(mysqli_query($conn, "INSERT INTO ingSuppliers (name,notes) VALUES ('$name','$description')")){
+		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Supplier '.$name.' added!</div>';
+	}
+	return;
+
+}elseif($_GET['supp'] == 'delete' && $_GET['ID']){
+	$ID = mysqli_real_escape_string($conn, $_GET['ID']);
+	$supplier = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingSuppliers WHERE id = '$ID'"));
+
+	if(mysqli_query($conn, "DELETE FROM ingSuppliers WHERE id = '$ID'")){
+		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Supplier <strong>'.$supplier['name'].'</strong> removed!</div>';
+	}
+	return;
 	
 }elseif($_GET['bottle']){
 	$value = mysqli_real_escape_string($conn, $_POST['value']);